pyproject.toml
[project]
name = "dhlab-sifi2000-lecture-2025"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Nasjonalbiblioteket - Seksjon for Språkbank og DH-lab"},
]
dependencies = ["datasets>=4.1.0", "transformers>=4.56.1", "pillow>=11.3.0", "httpx>=0.28.1", "torch>=2.7.1"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
